Once again refactoring: - new module pyrcn.base with:   - activation functions   - blocks
TODO: update all example scripts and notebooks!
