+++ pyrcn/extreme_learning_machine/tests/test_elm.py
@@ -11,7 +11,7 @@ from sklearn.model_selection import train_test_split
 from sklearn.model_selection import GridSearchCV
 from sklearn.pipeline import Pipeline, FeatureUnion
 
-from pyrcn.base import InputToNode
+from pyrcn.base.blocks import InputToNode
 from pyrcn.linear_model import IncrementalRegression
 from pyrcn.extreme_learning_machine import ELMClassifier, ELMRegressor
