+++ pyrcn/util/__init__.py
@@ -5,7 +5,7 @@ The :mod:`pyrcn.util` contains utilities for running, testing and analyzing the 
 # Author: Peter Steiner <peter.steiner@tu-dresden.de>, Michael Schindler <michael.schindler@maschindler.de>
 # License: BSD 3 clause
 
-from pyrcn.util._util import new_logger, tud_colors, get_mnist, argument_parser, export_ragged_time_series
+from pyrcn.util._util import new_logger, tud_colors, get_mnist, argument_parser, export_ragged_time_series, concatenate_sequences
 from pyrcn.util._feature_extractor import FeatureExtractor
 
 __all__ = ['new_logger',
@@ -13,5 +13,6 @@ __all__ = ['new_logger',
            'get_mnist',
            'argument_parser',
            'export_ragged_time_series',
-           'FeatureExtractor'
+           'FeatureExtractor',
+           'concatenate_sequences'
            ]
