--- pyrcn/utils/__init__.py
@@ -1,11 +1,0 @@
-"""
-The :mod:`pyrcn.util` contains utilities for running, testing and analyzing the reservoir computing modules
-"""
-
-# Author: Peter Steiner <peter.steiner@tu-dresden.de>
-# License: BSD 3 clause
-
-from pyrcn.utils._utils import concatenate_sequences
-
-__all__ = ['concatenate_sequences'
-           ]
