--- pyrcn/utils/tests/.gitignore
@@ -1,2 +1,0 @@
-*.log
-*.npz
